move animation to NotificationTest.tsx
