+++ packages/components/Notification/src/Notification.tsx
@@ -1,8 +1,7 @@
 /** @jsx withSlots */
-import { useLayoutEffect, useState } from 'react';
 import { notification, NotificationType, NotificationProps } from './Notification.types';
 import { Pressable } from '@fluentui-react-native/pressable';
-import { Animated, Easing, PressableProps, View, ViewStyle } from 'react-native';
+import { PressableProps, View, ViewStyle } from 'react-native';
 import { Icon } from '@fluentui-react-native/icon';
 import { Text } from '@fluentui-react-native/experimental-text';
 import { stylingSettings } from './Notification.styling';
@@ -55,58 +54,15 @@ export const Notification = compose<NotificationType>({
       const iconProps = createIconProps(icon);
       const notificationButtonProps = createNotificationButtonProps(userProps);
 
-      const [hidden, setHidden] = useState<boolean>(!visible);
-      const height = new Animated.Value(0);
-      const showDuration = isBar ? 300 : 600;
-      const elasticValue = isBar ? 0 : 1.5;
-
-      useLayoutEffect(() => {
-        if (visible) {
-          setHidden(false);
-          Animated.timing(height, {
-            toValue: -50,
-            duration: showDuration,
-            useNativeDriver: true,
-            easing: Easing.elastic(elasticValue),
-          }).start();
-        } else {
-          Animated.timing(height, {
-            toValue: 0,
-            duration: 250,
-            useNativeDriver: true,
-          }).start(() => {
-            setHidden(true);
-          });
-        }
-      }, [visible, height]);
-
-      const animatedViewProps = {
-        transform: [{ translateY: height }],
-      };
-
-      if (notificationButtonProps && Object.keys(notificationButtonProps).length === 1) {
-        notificationButtonProps.onClick = () => {
-          Animated.timing(height, {
-            toValue: 0,
-            duration: 250,
-            useNativeDriver: true,
-          }).start(() => {
-            setHidden(true);
-          });
-        };
-      }
-
-      return hidden ? null : (
-        <Animated.View style={[animatedViewProps]}>
-          <Slots.root {...mergedProps}>
-            {icon && <Slots.icon {...iconProps} />}
-            <Slots.contentContainer>
-              {title && <Slots.title>{title}</Slots.title>}
-              <Slots.message style={messageStyle}>{children}</Slots.message>
-            </Slots.contentContainer>
-            <Slots.action {...notificationButtonProps} />
-          </Slots.root>
-        </Animated.View>
+      return (
+        <Slots.root {...mergedProps}>
+          {icon && <Slots.icon {...iconProps} />}
+          <Slots.contentContainer>
+            {title && <Slots.title>{title}</Slots.title>}
+            <Slots.message style={messageStyle}>{children}</Slots.message>
+          </Slots.contentContainer>
+          <Slots.action {...notificationButtonProps} />
+        </Slots.root>
       );
     };
   },
